added currency formatting, added styling for safari browser

+++ src/main/java/org/perscholas/BankOfAntMerica/database/DAO/AccountTransactionDAO.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface AccountTransactionDAO extends JpaRepository<AccountTransaction, Integer> {
+public interface AccountTransactionDAO extends JpaRepository<AccountTransaction, Double> {
     List<AccountTransaction>findByBranchId(Integer id);
     List<AccountTransaction>findByAccountId(Integer id);
     AccountTransaction findAccountTransactionById(Integer id);
